+++ src/hooks/postMessage.test.ts
@@ -4,16 +4,21 @@ import { renderHook } from '@testing-library/react';
 import { v4 } from 'uuid';
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
-import { API_HOST, buildMockLocalContext } from '../../test/constants';
-import { mockHook, mockWindowForPostMessage, setUpTest } from '../../test/utils';
-import { defaultContextValue } from '../components/withContext';
-import { DEFAULT_CONTEXT, DEFAULT_LANG, DEFAULT_PERMISSION, MOCK_TOKEN } from '../config/constants';
+import { API_HOST, buildMockLocalContext } from '../../test/constants.js';
+import { mockHook, mockWindowForPostMessage, setUpTest } from '../../test/utils.js';
+import { defaultContextValue } from '../components/withContext.js';
+import {
+  DEFAULT_CONTEXT,
+  DEFAULT_LANG,
+  DEFAULT_PERMISSION,
+  MOCK_TOKEN,
+} from '../config/constants.js';
 import {
   MissingAppKeyError,
   MissingAppOriginError,
   MissingMessageChannelPortError,
-} from '../config/errors';
-import { AUTH_TOKEN_KEY, LOCAL_CONTEXT_KEY, buildPostMessageKeys } from '../config/keys';
+} from '../config/errors.js';
+import { AUTH_TOKEN_KEY, LOCAL_CONTEXT_KEY, buildPostMessageKeys } from '../config/keys.js';
 
 const mockItemId = 'mock-item-id';
 const POST_MESSAGE_KEYS = buildPostMessageKeys(mockItemId);
@@ -46,7 +51,7 @@ describe.skip('PostMessage Hooks', () => {
         const { data } = await mockHook({ hook, wrapper });
         expect(data).toEqual({
           apiHost: undefined, // @see LocalContext
-          memberId: undefined, // @see LocalContext
+          accountId: undefined, // @see LocalContext
           itemId: undefined, // @see LocalContext
           context: DEFAULT_CONTEXT,
           lang: DEFAULT_LANG,
@@ -67,7 +72,7 @@ describe.skip('PostMessage Hooks', () => {
           permission: 'write',
           lang: 'fr',
           itemId: v4(),
-          memberId: v4(),
+          accountId: v4(),
           settings: { some: 'value' },
           offline: true,
           dev: true,