+++ src/mockServer/msw/server.ts
@@ -1,8 +1,10 @@
+import { LocalContext } from '@graasp/sdk';
+
 import { SetupWorker, setupWorker } from 'msw/browser';
 
-import { Database, LocalContext } from '../../types';
-import { buildMockLocalContext } from '../fixtures';
-import { buildMSWMocks } from './handlers';
+import { Database } from '../../types.js';
+import { buildMockLocalContext } from '../fixtures.js';
+import { buildMSWMocks } from './handlers.js';
 
 /**
  * Creates and launches a mock server using MSW and IndexedDB
@@ -40,7 +42,7 @@ export const mockServiceWorkerServer = ({
       if (database.uploadedFiles?.length) {
         transaction.table('uploadedFiles').bulkAdd(database.uploadedFiles);
       }
-      transaction.table('appContext').add(fullAppContext, fullAppContext.memberId);
+      transaction.table('appContext').add(fullAppContext, fullAppContext.accountId);
     } else {
       console.debug('There was no data to populate the database');
     }