+++ src/mutations/appSetting.test.ts
@@ -12,16 +12,16 @@ import {
   UNAUTHORIZED_RESPONSE,
   buildAppSetting,
   buildMockLocalContext,
-} from '../../test/constants';
-import { mockMutation, setUpTest, waitForMutation } from '../../test/utils';
+} from '../../test/constants.js';
+import { mockMutation, setUpTest, waitForMutation } from '../../test/utils.js';
 import {
   buildDeleteAppSettingRoute,
   buildPatchAppSettingRoute,
   buildPostAppSettingRoute,
-} from '../api/routes';
-import { MOCK_TOKEN } from '../config/constants';
-import { AUTH_TOKEN_KEY, LOCAL_CONTEXT_KEY, appSettingKeys } from '../config/keys';
-import { patchAppSettingRoutine, postAppSettingRoutine } from '../routines';
+} from '../api/routes.js';
+import { MOCK_TOKEN } from '../config/constants.js';
+import { AUTH_TOKEN_KEY, LOCAL_CONTEXT_KEY, appSettingKeys } from '../config/keys.js';
+import { patchAppSettingRoutine, postAppSettingRoutine } from '../routines/index.js';
 
 const mockedNotifier = vi.fn();
 const { wrapper, queryClient, mutations } = setUpTest({
@@ -150,12 +150,12 @@ describe('App Settings Mutations', () => {
         // since the itemid is not defined, we do not check data for its key
       });
 
-      it('Throw if memberId is undefined', async () => {
+      it('Throw if accountId is undefined', async () => {
         // set necessary data
         queryClient.setQueryData(AUTH_TOKEN_KEY, MOCK_TOKEN);
         queryClient.setQueryData(LOCAL_CONTEXT_KEY, {
           ...buildMockLocalContext({ itemId }),
-          memberId: null,
+          accountId: null,
         });
         queryClient.setQueryData(key, initData);
 
@@ -342,12 +342,12 @@ describe('App Settings Mutations', () => {
         // since the itemId is not defined, we do not check data for its key
       });
 
-      it('Throw if memberId is undefined', async () => {
+      it('Throw if accountId is undefined', async () => {
         // set necessary data
         queryClient.setQueryData(AUTH_TOKEN_KEY, MOCK_TOKEN);
         queryClient.setQueryData(LOCAL_CONTEXT_KEY, {
           ...buildMockLocalContext({ itemId }),
-          memberId: null,
+          accountId: null,
         });
         queryClient.setQueryData(key, initData);
 
@@ -522,12 +522,12 @@ describe('App Settings Mutations', () => {
         // since the itemId is not defined, we do not check data for its key
       });
 
-      it('Throw if memberId is undefined', async () => {
+      it('Throw if accountId is undefined', async () => {
         // set necessary data
         queryClient.setQueryData(AUTH_TOKEN_KEY, MOCK_TOKEN);
         queryClient.setQueryData(LOCAL_CONTEXT_KEY, {
           ...buildMockLocalContext({ itemId }),
-          memberId: null,
+          accountId: null,
         });
 
         const initData = [toDelete];